requirements.txt
RPi.GPIO
spidev
Pillow
requests
python-dotenv
numpy
schedule
gpiozero
waveshare-epaper
pigpio
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client==2.182.0
Flask==3.0.3
